Bombs, animations of explosion, ability to destroy the walls
Its now possible to destroy walls by placing a bomb and waiting for it to explode. The check for the wall to be destroyed uses the vector2D created by the gamemanager, no need to do expensive collision checks because the grid has all the information necessary.
Its also worth pointing that wall the instantiation and removal of bombs and explosions are seem in the client as well, even though game is server authoritative.