PowerUp drops
Now its possible to pick powerUps when destroying a wall. The powerUp percentage rate is defined by the game controller, allowing holding more control over the drops, the powerUp is also set at the start of the game instead of when the block is destroyed. The client can also get powerUps.