Changed the movement to owner authorative
The main reason to this change is to prevent input lag, even using reconciliation to "fix" the lag delay I choose to adapt the game rules to this direction. Because the game is not meant to be played seriously and to be played with your friends hacking wont be a problem.